+++ lib/policies/msRestUserAgentPolicy.browser.ts
@@ -2,8 +2,6 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 import { MsRestUserAgentBase } from "./msRestUserAgentBase";
-import { HttpOperationResponse } from "../httpOperationResponse";
-import { WebResource } from "../webResource";
 import { RequestPolicy, RequestPolicyOptions } from "./requestPolicy";
 
 export const msRestUserAgentPolicy: any = () => {
@@ -11,15 +9,11 @@ export const msRestUserAgentPolicy: any = () => {
 };
 
 export class MsRestUserAgentPolicy extends MsRestUserAgentBase {
-
-  userAgentInfo: Array<string>;
-
-  constructor(nextPolicy: RequestPolicy, options: RequestPolicyOptions, userAgentInfo: Array<string>) {
-    super(nextPolicy, options);
-    this.userAgentInfo = userAgentInfo;
+  protected getUserAgentKey(): string {
+    return "x-ms-commandname";
   }
 
-  public sendRequest(webResource: WebResource): Promise<HttpOperationResponse> {
-    throw new Error("Method not implemented.");
+  constructor(nextPolicy: RequestPolicy, options: RequestPolicyOptions, userAgentInfo: Array<string>) {
+    super(nextPolicy, options, userAgentInfo);
   }
 }
